feat: add shortcut service
