+++ src/components/LayersPanel.vue
@@ -87,7 +87,7 @@ import blockIcons from '../image/layer_block';
 import { useService } from '../services';
 import { useContextMenuStore } from '../stores/contextMenu';
 
-const { viewport: viewportStore, nodeTree, nodes, pixels: pixelStore, output, keyboardEvent: keyboardEvents } = useStore();
+const { viewport: viewportStore, nodeTree, nodes, pixels: pixelStore, output } = useStore();
 const { layerPanel, layerQuery, nodeQuery, viewport, stageResize: stageResizeService, layerTool: layerSvc } = useService();
 const contextMenu = useContextMenuStore();
 
@@ -383,33 +383,6 @@ function deleteNode(id) {
     output.commit();
 }
 
-function deleteSelection() {
-    if (!nodeTree.selectedNodeCount) return;
-    output.setRollbackPoint();
-    const ids = nodeTree.selectedIds;
-    const lowermostTarget = nodeQuery.lowermost(ids);
-    const parentId = nodeQuery.parentOf(lowermostTarget);
-    const belowId = nodeQuery.below(lowermostTarget);
-    const removed = nodeTree.remove(ids);
-    nodes.remove(removed);
-    pixelStore.remove(removed);
-    let newSelect = null;
-    if (nodeTree.has(belowId)) {
-        newSelect = belowId;
-    } else {
-        const siblings = nodeQuery.childrenOf(parentId);
-        const lowermostSibling = nodeQuery.lowermost(siblings);
-        if (nodeTree.has(lowermostSibling)) {
-            newSelect = lowermostSibling;
-        } else if (nodeTree.has(parentId)) {
-            newSelect = parentId;
-        }
-    }
-    layerPanel.setRange(newSelect, newSelect);
-    if (newSelect) layerPanel.setScrollRule({ type: 'follow', target: newSelect });
-    output.commit();
-}
-
 function ensureBlockVisibility({
     type,
     target
@@ -480,63 +453,6 @@ function ensureBlockVisibility({
         behavior: 'smooth'
     });
 }
-
-  watch(() => keyboardEvents.recent.down, (downs) => {
-      for (const e of downs) {
-          const key = e.key;
-          const ctrl = e.ctrlKey || e.metaKey;
-          const shift = e.shiftKey;
-          switch (key) {
-              case 'ArrowUp':
-                  e.preventDefault();
-                  layerPanel.onArrowUp(shift, ctrl);
-                  break;
-              case 'ArrowDown':
-                  e.preventDefault();
-                  layerPanel.onArrowDown(shift, ctrl);
-                  break;
-              case 'Delete':
-              case 'Backspace':
-                  e.preventDefault();
-                  deleteSelection();
-                  break;
-              case 'Enter':
-                  if (!ctrl && !shift && nodeTree.selectedLayerCount === 1) {
-                      const selectedId = nodeTree.selectedLayerIds[0];
-                      const row = document.querySelector(`.layer[data-id="${selectedId}"] .nameText`);
-                      if (row) {
-                          e.preventDefault();
-                          row.dispatchEvent(new MouseEvent('dblclick', { bubbles: true }));
-                      }
-                  }
-                  break;
-              case 'Escape':
-                  if (output.hasPendingRollback) {
-                      e.preventDefault();
-                      output.rollbackPending();
-                  } else {
-                      nodeTree.clearSelection();
-                  }
-                  break;
-          }
-        if (ctrl) {
-            const lower = key.toLowerCase();
-            if (lower === 'a') {
-                e.preventDefault();
-                layerPanel.selectAll();
-            } else if (lower === 'g') {
-                e.preventDefault();
-                output.setRollbackPoint();
-                const ordered = nodeTree.orderedSelection;
-                nodeTree.replaceSelection(ordered);
-                const id = layerSvc.groupSelected();
-                layerPanel.setRange(id, id);
-                layerPanel.setScrollRule({ type: 'follow', target: id });
-                output.commit();
-            }
-        }
-      }
-  });
 
   watch(() => layerPanel.scrollRule, rule => nextTick(() => ensureBlockVisibility(rule)));
