+++ src/services/index.js
@@ -10,6 +10,7 @@ import { useStageResizeService } from './stageResize';
 import { useHamiltonianService } from './hamiltonian';
 import { useImageLoadService } from './imageLoad';
 import { useSettingsService } from './settings';
+import { useShortcutService } from './shortcut';
 
 export {
     useLayerPanelService,
@@ -30,7 +31,8 @@ export {
     useStageResizeService,
     useHamiltonianService,
     useImageLoadService,
-    useSettingsService
+    useSettingsService,
+    useShortcutService
 };
 
 export const useService = () => ({
@@ -54,5 +56,6 @@ export const useService = () => ({
     stageResize: useStageResizeService(),
     hamiltonian: useHamiltonianService(),
     imageLoad: useImageLoadService(),
-    settings: useSettingsService()
+    settings: useSettingsService(),
+    shortcut: useShortcutService()
 });
